Public API

StratIntervals.FourParBeta

StratIntervals.FourParBetaSampler

StratIntervals.RefOffExponential

StratIntervals.StratInterval

StratIntervals.ThreeParBeta

StratIntervals.ThreeParBetaSampler

Base.maximum

Base.maximum

Base.maximum

Base.minimum

Base.minimum

Base.minimum

Base.rand

Base.rand

Base.rand

Distributions.cdf

Distributions.cdf

Distributions.cdf

Distributions.insupport

Distributions.insupport

Distributions.insupport

Distributions.logpdf

Distributions.logpdf

Distributions.logpdf

Distributions.pdf

Distributions.pdf

Distributions.pdf

Distributions.pdf

Distributions.product_distribution

Statistics.quantile

Statistics.quantile

Statistics.quantile

StratIntervals.conflate

StratIntervals.conflate

StratIntervals.conflate

StratIntervals.quant

StratIntervals.quant

StratIntervals.sample_stratinterval

StratIntervals.sample_stratinterval

StratIntervals.tau_collection

StratIntervals.FourParBeta — Type

FourParBeta

Structure for constructing the PDF of stratigraphic interval estimation. It is a subtype of ContinuousUnivariateDitribution. This structure has four fields, corresponding to parameters: bounds on x [θ1,θ2] in years before present and the preservation parameters α, and β.

examples

using Distributions
rand(FourParBeta(5.0, 1.0, 1.0, 1.0)) # sample a random number from the FourParBeta with params θ1=5.0, θ2=1.0, α=1.0, β=1.0

StratIntervals.FourParBetaSampler — Type

FourParBetaSampler

This structure provides the parameters and function type for the FourParBeta, to be used with Distributions.rand. It is a subtype Sampleable of type Univariate, Continuous.

StratIntervals.RefOffExponential — Type

RefOffExponential

Structure for constructing the PDF of a reflected and offset exponential distribution. It is a subtype of ContinuousUnivariateDitribution. This structure has three fields, corresponding to parameters: theta, offset, and reflect.

examples

using Distributions
rand(RefOffExponential(1, 10.0, 1)) # sample a random number from the RefOffExponential with params θ=1.0, o=10.0, ρ=1

StratIntervals.StratInterval — Type

StratInterval(data, θ1_prior, θ2_prior, λ_prior)

This structure allows to input the data for estimation of stratigraphic intervals. Four attributes are available:

data, a Vector{Real} with the time occurrences, or a Vector{ContinuousUnivariateDistribution} with time intervals instead of point data

θ1_prior, a ContinuousUnivariateDistribution, specifying the prior on θ1,

θ2_prior, a ContinuousUnivariateDistribution, specifying the prior on θ2, and

λ_prior, a ContinuousUnivariateDistribution, specifying the prior on λ.

When sampling all these parameters, the priors need to be distributions. However, it is possible to fix parameters to concrete values by specifying a Float64 instead of a ContinuousUnivariateDistribution. Care must be taken to avoid fixing all of the parameters, at least one needs to be a ContinuousUnivariateDistribution.

Fixing a prior implies fixing its value in the model, so this needs to be used with extreme caution.

Fixing a prior essentially means calculating the posterior conditional to that observed value, and therefore does not correspond to the whole posterior. This is useful for simulation but is discouraged in empirical applications, no matter how certain we seem to be about the value for a given parameter.

An important exception to the rule above is when we have stratigraphic data for an extant lineage, in this case we can (and in fact should) fix θ2_prior to 0.0

examples

StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], Normal(10, 2), Exponential(1), Normal(0, 1)) # use priors for all parameters
StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], Normal(10, 2), 1.5, Normal(0, 1)) # fix θ2_prior to 1.5
StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], 9.8, 1.5, Normal(0, 1)) # fix θ2_prior to 1.5 and θ1_prior to 9.8
StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], 9.8, 1.5, 0.0) # this triggers an error
StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], Normal(10, 2), Exponential(1), 0.0) # fix λ_prior to 0.0
StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], Normal(10, 2), 0.0, Normal(0, 1)) # fix θ2_prior to 0.0 because the lineage is extant
StratInterval([Uniform(0, 2), Uniform(1, 2), Uniform(4, 6), Uniform(5, 9)], Normal(10, 2), Exponential(1), Normal(0, 1)) # intervals instead of point data

StratIntervals.ThreeParBeta — Type

ThreeParBeta

Structure for constructing the PDF of stratigraphic interval estimation. It is a subtype of ContinuousUnivariateDitribution. This structure has three fields, corresponding to parameters: bounds on x [θ1,θ2] in years before present and the preservation parameter λ.

examples

using Distributions
rand(ThreeParBeta(5.0, 1.0, 0.0)) # sample a random number from the ThreeParBeta with params θ1=5.0, θ2=1.0, λ=0.0

StratIntervals.ThreeParBetaSampler — Type

ThreeParBetaSampler

This structure provides the parameters and function type for the ThreeParBeta, to be used with Distributions.rand. It is a subtype Sampleable of type Univariate, Continuous.

Base.maximum — Method

Distributions.maximum(d::FourParBeta)

This method extends the maximum function for the type FourParBeta. See Distributions.maximum. This function returns the numerical maximum bound in the support of the FourParBeta distribution. It is the parameter θ1.

examples

using Distributions
maximum(FourParBeta(5.0, 1.0, 1.0, 1.0))

Base.maximum — Method

Distributions.maximum(d::RefOffExponential)

This method extends the maximum function for the type RefOffExponential. See Distributions.maximum. This function returns the numerical maximum bound in the support of the RefOffExponential distribution. It depends on whether it is reflected and offset or not.

examples

using Distributions
minimum(RefOffExponential(1, 10.0, 1))

Base.maximum — Method

Distributions.maximum(d::ThreeParBeta)

This method extends the maximum function for the type ThreeParBeta. See Distributions.maximum. This function returns the numerical maximum bound in the support of the ThreeParBeta distribution. It is the parameter θ1.

examples

using Distributions
maximum(ThreeParBeta(5.0, 1.0, 0.0))

Base.minimum — Method

Distributions.minimum(d::FourParBeta)

This method extends the minimum function for the type FourParBeta. See Distributions.minimum. This function returns the numerical minimum bound in the support of the FourParBeta distribution. It is the parameter θ2.

examples

using Distributions
minimum(FourParBeta(5.0, 1.0, 1.0, 1.0))

Base.minimum — Method

Distributions.minimum(d::RefOffExponential)

This method extends the minimum function for the type RefOffExponential. See Distributions.minimum. This function returns the numerical minimum bound in the support of the RefOffExponential distribution. It depends on whether it is reflected and offset or not.

examples

using Distributions
minimum(RefOffExponential(1, 10.0, 1))

Base.minimum — Method

Distributions.minimum(d::ThreeParBeta)

This method extends the minimum function for the type ThreeParBeta. See Distributions.minimum. This function returns the numerical minimum bound in the support of the ThreeParBeta distribution. It is the parameter θ2.

examples

using Distributions
minimum(ThreeParBeta(5.0, 1.0, 0.0))

Base.rand — Method

Base.rand(rng::AbstractRNG, d::FourParBetaSampler)

This method extends the random sampler for the FourParBeta. It will return a single random Float64 from the FourParBeta when called with only the distribution specification, or a vector of Float64 when specifying how many numbers to sample.

examples

using Distributions
rand(FourParBeta(5.0, 1.0, 1.0, 1.0)) # return a single random number
rand(FourParBeta(5.0, 1.0, 1.0, 1.0), 10) # return a vector of 10 random numbers

Base.rand — Method

Base.rand(rng::AbstractRNG, d::RefOffExponential)

This method extends the random sampler for the RefOffExponential. It will return a single random Float64 from the RefOffExponential when called with only the distribution specification, or a vector of Float64 when specifying how many numbers to sample.

examples

using Distributions
rand(RefOffExponential(1, 10.0, 1)) # return a single random number
rand(RefOffExponential(1, 10.0, 1), 10) # return a vector of 10 random numbers

Base.rand — Method

Base.rand(rng::AbstractRNG, d::ThreeParBetaSampler)

This method extends the random sampler for the ThreeParBeta. It will return a single random Float64 from the ThreeParBeta when called with only the distribution specification, or a vector of Float64 when specifying how many numbers to sample.

examples

using Distributions
rand(ThreeParBeta(5.0, 1.0, 0.0)) # return a single random number
rand(ThreeParBeta(5.0, 1.0, 0.0), 10) # return a vector of 10 random numbers

Distributions.cdf — Method

Distributions.cdf(d::FourParBeta, τ::Real)

This method extends the cdf function for the type FourParBeta. See Distributions.cdf. The function returns the value of the cdf at a given value of τ.

examples

using Distributions
cdf(FourParBeta(5.0, 1.0, 1.0, 1.0), 2.5)

Distributions.cdf — Method

Distributions.cdf(d::RefOffExponential, x::Real)

This method extends the cdf function for the type RefOffExponential. See Distributions.cdf. The function returns the value of the cdf at a given value of x.

examples

using Distributions
cdf(RefOffExponential(1, 10.0, 1), 2.5)

Distributions.cdf — Method

Distributions.cdf(d::ThreeParBeta, τ::Real)

This method extends the cdf function for the type ThreeParBeta. See Distributions.cdf. The function returns the value of the cdf at a given value of τ.

examples

using Distributions
cdf(ThreeParBeta(5.0, 1.0, 0.0), 2.5)

Distributions.insupport — Method

Distributions.insupport(d::FourParBeta)

This method extends the insupport function for the type FourParBeta. See Distributions.insupport. This function returns a Boolean being true if τ is in the support [θ1,θ2] of the FourParBeta distribution and false otherwise.

examples

using Distributions
insupport(FourParBeta(5.0, 1.0, 1.0, 1.0), 0.5) # false
insupport(FourParBeta(5.0, 1.0, 1.0, 1.0), 3.2) # true

Distributions.insupport — Method

Distributions.insupport(d::RefOffExponential, x::Real)

This method extends the insupport function for the type RefOffExponential. See Distributions.insupport. This function returns a Boolean being true if x is in the support (-Inf,offset] when ρ = -1, or [offset,Inf) when standard, possible offset, and false otherwise.

examples

using Distributions
insupport(RefOffExponential(1, 10.0, 1), 0.5) # false
insupport(RefOffExponential(1, 10.0, 1), 12.0) # true

Distributions.insupport — Method

Distributions.insupport(d::ThreeParBeta)

This method extends the insupport function for the type ThreeParBeta. See Distributions.insupport. This function returns a Boolean being true if τ is in the support [θ1,θ2] of the ThreeParBeta distribution and false otherwise.

examples

using Distributions
insupport(ThreeParBeta(5.0, 1.0, 0.0), 0.5) # false
insupport(ThreeParBeta(5.0, 1.0, 0.0), 3.2) # true

Distributions.logpdf — Method

Distributions.logpdf(d::FourParBeta, τ::Real)

This method extends the logpdf function for the type FourParBeta. See Distributions.logpdf. The function returns the value of the logpdf at a given value of τ.

examples

using Distributions
logpdf(FourParBeta(5.0, 1.0, 1.0, 1.0), 2.5)

Distributions.logpdf — Method

Distributions.logpdf(d::RefOffExponential, x::Real)

This method extends the logpdf function for the type RefOffExponential. See Distributions.logpdf. The function returns the value of the logpdf at a given value of x.

examples

using Distributions
logpdf(RefOffExponential(1, 10.0, 1), 2.5)

Distributions.logpdf — Method

Distributions.logpdf(d::ThreeParBeta, τ::Real)

This method extends the logpdf function for the type ThreeParBeta. See Distributions.logpdf. The function returns the value of the logpdf at a given value of τ.

examples

using Distributions
logpdf(ThreeParBeta(5.0, 1.0, 0.0), 2.5)

Distributions.pdf — Method

Distributions.pdf(d::FouParBeta, τ::Real)

This method extends the pdf function for the type FourParBeta so that the PDF is calculated given the parameters θ1, θ2, α, and β. The function returns the value of the PDF at a given value of τ.

examples

using Distributions
pdf(FourParBeta(5.0, 1.0, 1.0, 1.0), 2.5)

Distributions.pdf — Method

Distributions.pdf(d::Product, x::Float64)

Method for specifying the product of an arbitrary vector of pdfs to be evaluated at the same x value.

Note: This method may need to be changed when Distributions deprecate Product. See product_distribution(dists::Vector{Normal{Float64}}).

examples

using Distributions
using Random
Random.seed!(4);
distribs = [Normal(0, 1), Normal(1, 1), Normal(2, 1)]
pdf(product_distribution(distribs), 1.5) # evaluate the product of three normals at x=1.5

Distributions.pdf — Method

Distributions.pdf(d::RefOffExponential, x::Real)

This method extends the pdf for the type RefOffExponential so that the PDF is calculated given the parameters θ, o, and ρ. The function returns the value of the PDF at a given value of x.

examples

using Distributions
pdf(RefOffExponential(1, 10.0, 1), 2.5)

Distributions.pdf — Method

Distributions.pdf(d::ThreeParBeta, τ::Real)

This method extends the pdf function for the type ThreeParBeta so that the PDF is calculated given the parameters θ1, θ2 and λ. The function returns the value of the PDF at a given value of τ.

examples

using Distributions
pdf(ThreeParBeta(5.0, 1.0, 0.0), 2.5)

Distributions.product_distribution — Method

Distributions.product_distribution(dists::Vector{Normal{Float64}})

Method for making product of Normals to behave as the generic one. The specialised one returning a MvNormal generates issues during conflation when all distribs are normal.

The methods associated to Product are going to be deprecated from Distributions. Prepare to modify this method accordingly.

Statistics.quantile — Method

Distributions.quantile(d::FourParBeta, q::Real)

This method extends the quantile function for the type FourParBeta. See Distributions.quantile and Distributions.quantile(d::UnivariateDistribution, q::Real).

examples

using Distributions
quantile(FourParBeta(5.0, 1.0, 1.0, 1.0), 0.5) # calculate the median of the distribution

Statistics.quantile — Method

Distributions.quantile(d::RefOffExponential, q::Real)

This method extends the quantile function for the type RefOffExponential. See Distributions.quantile and Distributions.quantile(d::UnivariateDistribution, q::Real).

examples

using Distributions
quantile(RefOffExponential(1, 10.0, 1), 0.5) # calculate the median of the distribution

Statistics.quantile — Method

Distributions.quantile(d::ThreeParBeta, q::Real)

This method extends the quantile function for the type ThreeParBeta. See Distributions.quantile and Distributions.quantile(d::UnivariateDistribution, q::Real).

examples

using Distributions
quantile(ThreeParBeta(5.0, 1.0, 0.0), 0.5) # calculate the median of the distribution

StratIntervals.conflate — Method

conflate(d::Array{Float64, 2}, x::Float64)

Conflation method for a matrix of floats, which are coming from a posterior predictive analysis, one column per stratigraphic interval.

It takes a matrix of empirical values to smooth, column-wise, and then apply the KDE and interpolation on each column.

Then it applies the conflation to these empirical pdfs and evaluate it at x.

StratIntervals.conflate — Method

conflate(d::Product, x::Float64)

Conflation method for a product distribution, which are coming from calling pdf on a vector of pdfs.

It takes a vector of pdfs and conflate then evaluating the resulting pdf at x.

For conflation of empirical distributions see conflate(d::Array{Float64, 2}, x::Float64)@ref

StratIntervals.conflate — Method

conflate(d::Vector{ContinuousUnivariateDitribution}, x::Float64)

Conflation method for a vector of distributions.

It takes a vector of pdfs and conflate then evaluating the resulting pdf at x.

This is a convenience method for allowing either input as vector or product_distribution

StratIntervals.quant — Method

quant(taus::Array{Float64, 2}, p::Float64, lower_tail::Bool, min::Float64, max::Float64)

A function for calculating the quantile for a given probability value on a conflated distribution. Whether to calculate lower or upper tail is available. The method uses constrained numerical optimisation, and therefore a minimum and maximum need to be specified.

taus a matrix of stacked empirical distributions, one distribution per column p the desired probability lower_tail a boolean defining whether we want to calculate the probability from -Inf to quant, if false it calculates the prob from quant to Inf min minimum value for the box optimisation, normally 0.0 max maximum value for the box optimisation, this should be something sensible outside of the bulk of the conflated PDF

examples

using Random
using Distributions
using StatsPlots
using StratIntervals
Random.seed!(1507)

#empirical1
empirical1_vec = rand(Normal(0, 1), 20)
#empirical2
empirical2_vec = rand(Uniform(1, 2), 20)
# stack them together in a matrix, column-wise
empirical_matrix = [empirical1_vec empirical2_vec]

xx = -5:0.01:5
yy = map(x -> conflate(empirical_matrix, x), xx);

plot(xx, yy, label="Conflation")
density!(empirical1_vec, label="Empirical 1")
density!(empirical2_vec, label="Empirical 2")

quant(empirical_matrix, 0.025, true, -5.0, 5.0)
quant(empirical_matrix, 0.975, true, -5.0, 5.0)
quant(empirical_matrix, 0.025, false, -5.0, 5.0)
quant(empirical_matrix, 0.5, false, -5.0, 5.0)

StratIntervals.quant — Method

quant(taus::Vector{Any}, p::Float64, lower_tail::Bool, min::Float64, max::Float64)

A function for calculating the quantile for a given probability value on a conflated distribution. Whether to calculate lower or upper tail is available. The method uses constrained numerical optimisation, and therefore a minimum and maximum need to be specified.

taus a vector which comes from calling samplestratinterval(vector{StratInterval}) p the desired probability lowertail a boolean defining whether we want to calculate the probability from -Inf to quant, if false it calculates the prob from quant to Inf min minimum value for the box optimisation, normally 0.0 max maximum value for the box optimisation, this should be something sensible outside of the bulk of the conflated PDF

examples

using Random
using Distributions
using StatsPlots
using StratIntervals
using Turing
Random.seed!(1507)

setprogress!(false)

ndata = 100;
iters = 10000;

# first interval
true_lambda_1 = 0;
true_theta2_1 = 10;
true_theta1_1 = 22;
data_1 = rand(ThreeParBeta(true_theta1_1, true_theta2_1, true_lambda_1), ndata);

# second interval
true_lambda_2 = 1;
true_theta2_2 = 15;
true_theta1_2 = 30;
data_2 = rand(ThreeParBeta(true_theta1_2, true_theta2_2, true_lambda_2), ndata);

# specifying each StratInterval object
interval_1 = StratInterval(data_1, Normal(22, 3), Exponential(10), Normal(0,3));
interval_2 = StratInterval(data_2, Normal(30, 3), Exponential(15), Normal(1,3));

# construct the vector of StratIntervals
vecinterval = [interval_1, interval_2];

# MCMC sampling and posterior predictive
mystratint_postpredict_vec = sample_stratinterval(vecinterval, iters, NUTS(), false, true);

# now calculate the conflation of posterior predictives for the collection interval_1, interval_2

xx = 0:0.01:40;
yy = map(x -> tau_collection(mystratint_postpredict_vec, x), xx);

plot(xx, yy, label="Conflation")
density!(mystratint_postpredict_vec[1][2], label="Postpredictive 1")
density!(mystratint_postpredict_vec[2][2], label="Postpredictive 2")

quant(mystratint_postpredict_vec, 0.025, true, 0.0, 800.0)
quant(mystratint_postpredict_vec, 0.975, true, 0.0, 800.0)
quant(mystratint_postpredict_vec, 0.025, false, 0.0, 800.0)
quant(mystratint_postpredict_vec, 0.5, false, 0.0, 800.0)

StratIntervals.sample_stratinterval — Method

sample_stratinterval(data_priors::StratInterval, iters, sampler, prior, postpredict)

This function samples the posterior distributions of the stratigraphic interval using MCMC.

Input are the data and priors as constructed using StratInterval(). iters is the number of iterations during sampling, sampler is the kind of sampler to use, e.g., HMC(epsilon, tau) or NUTS(), whether to sample from the prior, and finally the last argument specifies whether we want to calculate the posterior predictive distribution of τ.

examples

using Distributions
using Turing
using StratIntervals
my_strinterval = StratInterval([2.0, 3.1, 3.2, 4.6, 6.77], Normal(10, 2), Exponential(1), Normal(0, 1))
sample_stratinterval(my_strinterval, 10000, NUTS(), true, false) # sample from prior
sample_stratinterval(my_strinterval, 10000, NUTS(), false, false) # sample from posterior
sample_stratinterval(my_strinterval, 10000, NUTS(), false, true) # sample from posterior and calculate posterior predictive of tau

StratIntervals.sample_stratinterval — Method

sample_stratinterval(data_priors::Vector{StratInterval}, iters, sampler, prior, postpredict)

This function samples the posterior distributions of a vector of stratigraphic intervals using MCMC.

Input data is a vector of type StratInterval. iters is the number of iterations during sampling, sampler is the kind of sampler to use, e.g., HMC(epsilon, tau) or NUTS(), whether to sample from the prior, and finally the last argument specifies whether we want to calculate the posterior predictive distribution of τ.

examples

using Distributions
using Turing
ndata = 100
iters = 10000
# first interval
true_lambda_1 = 0
true_theta2_1 = 10
true_theta1_1 = 22
data_1 = rand(ThreeParBeta(true_theta1_1, true_theta2_1, true_lambda_1), ndata)
# second interval
true_lambda_2 = 1
true_theta2_2 = 15
true_theta1_2 = 30
data_2 = rand(ThreeParBeta(true_theta1_2, true_theta2_2, true_lambda_2), ndata)
interval_1 = StratInterval(data_1, Normal(22, 3), Exponential(10), Normal(0,3))
interval_2 = StratInterval(data_2, Normal(30, 3), Exponential(15), Normal(1,3))
# construct the vector of StratIntervals
vecinterval = [interval_1, interval_2]
# MCMC sampling
mystratint_mcmc_vec = sample_stratinterval(vecinterval, iters, NUTS(), false, false)
# MCMC sampling and posterior predictive
mystratint_postpredict_vec = sample_stratinterval(vecinterval, iters, NUTS(), false, true)

StratIntervals.tau_collection — Method

tau_collection(taus, x::Float64)

Calculation of the pdf for tau as a conflation of StratIntervals.

This uses the output of sample_interval with a Vector{StratInterval} and where postpredict is true.

The value of x is the evaluated using the conflation of the collection of stratigraphic intervals, that is, the tau for the co-occurrence of StratIntervals.